flask/requirements.txt
flask==2.1.2
redis==4.3.1
Cmake==3.22.4
PyMySQL==1.0.2
cryptography==37.0.2
Werkzeug==2.1.2
mysql-connector-python